% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReGIFT-package.R
\docType{package}
\name{ReGIFT-package}
\alias{ReGIFT}
\alias{ReGIFT-package}
\title{ReGIFT: Replicate-Guided Matrix Factorization}
\description{
ReGIFT estimates condition-response programs that generalize across
biological replicates while separating donor-specific deviations and
technical variation. Use [regift()] for the standard user workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/haohaostats/ReGIFT}
  \item Report bugs at \url{https://github.com/haohaostats/ReGIFT/issues}
}

}
\author{
\strong{Maintainer}: ReGIFT Development Team \email{haohaostats@users.noreply.github.com}

}
\keyword{internal}
