% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{regift_predict_response}
\alias{regift_predict_response}
\title{Return population response matrices for fitted cells}
\usage{
regift_predict_response(object)
}
\arguments{
\item{object}{A regift_fit object.}
}
\description{
Return population response matrices for fitted cells
}
